feat: Add loading state to Store screen
This commit adds a loading state to the Store screen in order to provide visual feedback to the user while the transaction status data is being fetched from the server. The loading state is set to true initially and is set to false once the data is successfully retrieved or an error occurs. This enhancement improves the user experience by indicating that the data is being loaded and preventing any confusion or frustration caused by a blank screen.