feat: Update ToShip screen to use useFocusEffect for fetching data
This commit updates the ToShip screen to use the `useFocusEffect` hook from `@react-navigation/native` for fetching data. Previously, the `useEffect` hook was used, which caused the data to be fetched even when the screen was not in focus. By switching to `useFocusEffect`, the data will only be fetched when the screen is in focus, improving performance and reducing unnecessary API calls.