feat: Add functionality to send messages via SMS
The code changes in the SelectedStore.jsx file add functionality to send messages via SMS. This commit includes the addition of the Linking and Alert components, as well as the implementation of the handleMessage function. Users can now send messages to the shop owner by tapping the "Chat" button. This enhancement improves the communication experience between users and shop owners.