refactor: Update UserController to include search and get all users functionality
This commit refactors the UserController in the Users.js file to include new functionality for searching users and retrieving all users. The code changes add the SearchUsers and GetAllUsers methods, which enable searching for users based on a search ID and retrieving all users from the database, respectively. This enhancement improves the user management capabilities of the application.