refactor: Add profilePhoto field to UserSchema in Users.js
The code changes in Users.js add a new field called `profilePhoto` to the `UserSchema` in the `Users` model. This field is required and will store the URL of the user's profile photo. This update ensures that the user's profile photo can be properly stored and retrieved in the application.