refactor: Update UserController to handle profile photo update
The code changes in Users.js update the UserController to handle the update of the user's profile photo. Two new methods, UpdateProfilePhoto and UpdateAccountDetails, are added to handle the profile photo and account details updates respectively. This enhancement improves the functionality of the UserController by providing the ability to update the user's profile photo and account details, enhancing the user experience by allowing for personalized profile images and updated account information.