feat: Add IsDelivery parameter to SelectedItem endpoint
The code changes in `Products.js` add the `IsDelivery` parameter to the `SelectedItem` endpoint in the `ProductController`. This parameter is used to determine if the product should be delivered or not. The commit message follows the established convention of prefixing the commit type (`feat`) followed by a concise and descriptive message.