refactor: Add Employee page and data grid component
Add a new Employee page and a data grid component to display a list of employees. The page includes a search input field, an "Add Employee" button, and a data grid that fetches and displays employee data from the server. This change improves the user interface and functionality of the application.